Updates ffmpeg based on Vita3K repo to 5.1.2.

The new ffmpeg version is more optimized overall, so this is actually useful to performance during de/encoding.
Here some random timings of mostly the same scenes on repeat:
<details>
  <summary>Encoding (720p) ~20% faster</summary>
  
Old:
E Media: average: 5008.96 us
E Media: average: 5008.11 us
E Media: average: 5020.52 us
E Media: average: 5576.87 us
E Media: average: 6629.70 us
E Media: average: 6560.65 us

New:
E Media: average: 4092.24 us
E Media: average: 4101.77 us
E Media: average: 4087.86 us
E Media: average: 4050.64 us
E Media: average: 5006.84 us
E Media: average: 5860.35 us
</details>
<details>
  <summary>Encoding (4K) ~20% faster</summary>

4K Old:
E Media: average: 27370.50 us
E Media: average: 27518.19 us
E Media: average: 27480.09 us
E Media: average: 27321.86 us

4K New:
E Media: average: 21630.02 us
E Media: average: 21576.04 us
E Media: average: 21503.02 us
E Media: average: 21505.84 us
</details>
<details>
  <summary>Decoding (720p) ~10% faster</summary>
  
Old:
E cellVdec: average: 2724.74 us
E cellVdec: average: 2601.69 us
E cellVdec: average: 2631.09 us
E cellVdec: average: 2621.00 us
E cellVdec: average: 2594.64 us
E cellVdec: average: 2601.26 us
E cellVdec: average: 2602.29 us

New:
E cellVdec: average: 2476.39 us
E cellVdec: average: 2348.26 us
E cellVdec: average: 2372.78 us
E cellVdec: average: 2334.19 us
E cellVdec: average: 2346.55 us
E cellVdec: average: 2312.74 us
E cellVdec: average: 2347.38 us
</details>

The new commit drops the source and uses CI (see .github) and a patch with the options to build the libs.
The artifacts are created automatically on our fork in the github actions section.
The codecs seem to be the same still:

![image](https://github.com/RPCS3/rpcs3/assets/23019877/541b509d-9e43-46de-82f0-f64e542e3bef)