@DarthMDev 2023-06-26
Can this be updated again? A recent change was pushed to the ffmpeg repo to have macOS prebuilt arm64 libraries, and it would be helpful for #14028