@DarthMDev 2023-07-03
seems to be an issue with the macOS minimum version with the newer ffmpeg update

(was built for newer macOS version (12.6) than being linked (12.0))