@DarthMDev 2023-07-03
> seems to be an issue with the macOS minimum version with the newer ffmpeg update
> 
> (was built for newer macOS version (12.6) than being linked (12.0))
> 
> Maybe if we change CMAKE_OSX_DEPLOYMENT_TARGET to 12.6?

I made a pr to address this on your fork https://github.com/Megamouse/rpcs3/pull/15