@shinra-electric 2023-07-03
> seems to be an issue with the macOS minimum version with the newer ffmpeg update
> 
> (was built for newer macOS version (12.6) than being linked (12.0))
> 
> Maybe if we change CMAKE_OSX_DEPLOYMENT_TARGET to 12.6?

I think bumping the Cirrus image from Monterey (12) to Ventura (13) might help.