@Megamouse 2023-07-03
This PR is nowhere near being considered for merge.
This won't happen until Ubuntu 18.04 is dropped.
So all of these changes are not really worth spending time on right now