@Megamouse 2023-07-03
Also, The libs are built with macOs latest, as proof of concept.
There is not reason to bump the macOs version in this PR.
That should happen in a seperate PR if even needed.